The RED memory isn't used on this B/W display, no need to keep updating it
This code appears to be leftover from some other chipset which let you
use the extra "RED" bit memory to do double buffering of update data.
As far as I can tell from the SSD1681 datasheet, the driver for this
display doesn't have any way to do that: instead (Table 6-5 in the
datasheet) LUT2 = LUT0 and LUT3 = LUT1 so the red data bits are effectively
ignored.  Add a bit of code to set up 0x21 (RAM content option) to
*doubly ensure* that RED data bits are ignored.  Then we can stop
writing them.